BI-4121 Moving objections SDk into modules folder
